pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
